fix all the paths
the default layout of a django app is weird. I don't understand why we're still doing it this way, but three nested directories with the same name smells funky.

this flattens it one layer and brings it into conventional agreement with heroku's example django app and most other django apps I've ever worked with